refactor: moves files and formats files correct
