+++ src/rss/feed/YouTube.tsx
@@ -1,12 +1,7 @@
-
 import { YoutubeVideoItem } from '../index';
 import style from './feed.module.css';
 
-export default function YouTube({
-  item,
-}: {
-  item: YoutubeVideoItem;
-}) {
+export default function YouTube({ item }: { item: YoutubeVideoItem }) {
   return (
     <div className={`${style.card} ${style['card-youtube']}`}>
       <figure className={style.media}>
