+++ src/rss/feed/feed.module.css
@@ -1,3 +1,17 @@
+.feed {
+  display: grid;
+  grid-gap: 1.5rem;
+}
+@media (max-width: 600px) {
+  .feed {
+    padding: 1rem;
+  }
+}
+
+.feed__head {
+  padding: 3rem 2rem;
+}
+
 .title {
   margin-bottom: 0;
 }
